1.5.0: improve code organization and types
