+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "client-web-storage",
   "description": "Web browser (client) storage interface for IndexedDB, WebSQL, LocalStorage, and in memory data with Schema validation.",
-  "version": "1.4.5",
+  "version": "1.5.0",
   "main": "dist/",
   "repository": {
     "url": "https://github.com/beforesemicolon/client-web-storage",
@@ -23,7 +23,7 @@
     "build:client-local": "esbuild src/client.ts --bundle --target=es2020 --outfile=dist/client-web-storage.min.js",
     "build": "tsc --noEmit && rm -rf dist && npm run build:client && tsc",
     "build:local": "tsc --noEmit && rm -rf dist && npm run build:client-local && tsc",
-    "local": "nodemon --watch src -e ts --exec npm run build:local",
+    "local": "nodemon --watch src -e ts --exec 'npm run build:local && npm pack'",
     "test": "jest --coverage",
     "test:watch": "jest --watch"
   },
