+++ src/SchemaId.ts
@@ -1,0 +1,11 @@
+import {generateUUID} from "./utils/generate-uuid";
+
+export class SchemaId {
+	value = (() => {
+		try {
+			return crypto.randomUUID();
+		} catch (e) {
+			return generateUUID();
+		}
+	})();
+}
