+++ src/client.ts
@@ -1,4 +1,6 @@
-import {Schema, SchemaValue, SchemaId} from './Schema';
+import {Schema} from './Schema';
+import {SchemaId} from './SchemaId';
+import {SchemaValue} from './SchemaValue';
 import {ClientStore} from './ClientStore';
 
 // @ts-ignore
