+++ src/index.ts
@@ -1,2 +1,5 @@
-export {Schema, SchemaValue, SchemaId} from './Schema';
+export * from './types';
+export {Schema} from './Schema';
+export {SchemaId} from './SchemaId';
+export {SchemaValue} from './SchemaValue';
 export {ClientStore} from './ClientStore';
